+++ tests/cloudpickle_test.py
@@ -13,6 +13,7 @@ from operator import itemgetter, attrgetter
 import pickletools
 import platform
 import random
+import re
 import shutil
 import subprocess
 import sys
@@ -45,7 +46,10 @@ except ImportError:
 
 import cloudpickle
 from cloudpickle.compat import pickle
-from cloudpickle.cloudpickle import _is_importable
+from cloudpickle import register_pickle_by_value
+from cloudpickle import unregister_pickle_by_value
+from cloudpickle import list_registry_pickle_by_value
+from cloudpickle.cloudpickle import _should_pickle_by_reference
 from cloudpickle.cloudpickle import _make_empty_cell, cell_set
 from cloudpickle.cloudpickle import _extract_class_dict, _whichmodule
 from cloudpickle.cloudpickle import _lookup_module_and_qualname
@@ -81,6 +85,14 @@ def pickle_depickle(obj, protocol=cloudpickle.DEFAULT_PROTOCOL):
 def _escape(raw_filepath):
     # Ugly hack to embed filepaths in code templates for windows
     return raw_filepath.replace("\\", r"\\\\")
+
+
+def _maybe_remove(list_, item):
+    try:
+        list_.remove(item)
+    except ValueError:
+        pass
+    return list_
 
 
 def test_extract_class_dict():
@@ -732,24 +744,24 @@ class CloudPickleTest(unittest.TestCase):
         import distutils
         import distutils.ccompiler
 
-        assert _is_importable(pickle)
-        assert _is_importable(os.path)  # fake (aliased) module
-        assert _is_importable(distutils)  # package
-        assert _is_importable(distutils.ccompiler)  # module in package
+        assert _should_pickle_by_reference(pickle)
+        assert _should_pickle_by_reference(os.path)  # fake (aliased) module
+        assert _should_pickle_by_reference(distutils)  # package
+        assert _should_pickle_by_reference(distutils.ccompiler)  # module in package
 
         dynamic_module = types.ModuleType('dynamic_module')
-        assert not _is_importable(dynamic_module)
+        assert not _should_pickle_by_reference(dynamic_module)
 
         if platform.python_implementation() == 'PyPy':
             import _codecs
-            assert _is_importable(_codecs)
+            assert _should_pickle_by_reference(_codecs)
 
         # #354: Check that modules created dynamically during the import of
         # their parent modules are considered importable by cloudpickle.
         # See the mod_with_dynamic_submodule documentation for more
         # details of this use case.
         import _cloudpickle_testpkg.mod.dynamic_submodule as m
-        assert _is_importable(m)
+        assert _should_pickle_by_reference(m)
         assert pickle_depickle(m, protocol=self.protocol) is m
 
         # Check for similar behavior for a module that cannot be imported by
@@ -757,14 +769,14 @@ class CloudPickleTest(unittest.TestCase):
         from _cloudpickle_testpkg.mod import dynamic_submodule_two as m2
         # Note: import _cloudpickle_testpkg.mod.dynamic_submodule_two as m2
         # works only for Python 3.7+
-        assert _is_importable(m2)
+        assert _should_pickle_by_reference(m2)
         assert pickle_depickle(m2, protocol=self.protocol) is m2
 
         # Submodule_three is a dynamic module only importable via module lookup
         with pytest.raises(ImportError):
             import _cloudpickle_testpkg.mod.submodule_three  # noqa
         from _cloudpickle_testpkg.mod import submodule_three as m3
-        assert not _is_importable(m3)
+        assert not _should_pickle_by_reference(m3)
 
         # This module cannot be pickled using attribute lookup (as it does not
         # have a `__module__` attribute like classes and functions.
@@ -776,12 +788,12 @@ class CloudPickleTest(unittest.TestCase):
         # Do the same for an importable dynamic submodule inside a dynamic
         # module inside a file-backed module.
         import _cloudpickle_testpkg.mod.dynamic_submodule.dynamic_subsubmodule as sm  # noqa
-        assert _is_importable(sm)
+        assert _should_pickle_by_reference(sm)
         assert pickle_depickle(sm, protocol=self.protocol) is sm
 
         expected = "cannot check importability of object instances"
         with pytest.raises(TypeError, match=expected):
-            _is_importable(object())
+            _should_pickle_by_reference(object())
 
     def test_Ellipsis(self):
         self.assertEqual(Ellipsis,
@@ -2351,6 +2363,227 @@ class CloudPickleTest(unittest.TestCase):
         o = MyClass()
         pickle_depickle(o, protocol=self.protocol)
 
+    def test_pickle_constructs_from_module_registered_for_pickling_by_value(self):  # noqa
+        _prev_sys_path = sys.path.copy()
+        try:
+            # We simulate an interactive session that:
+            # - we start from the /path/to/cloudpickle/tests directory, where a
+            #   local .py file (mock_local_file) is located.
+            # - uses constructs from mock_local_file in remote workers that do
+            #   not have access to this file. This situation is
+            #   the justification behind the
+            #   (un)register_pickle_by_value(module) api that cloudpickle
+            #   exposes.
+            _mock_interactive_session_cwd = os.path.dirname(__file__)
+
+            # First, remove sys.path entries that could point to
+            # /path/to/cloudpickle/tests and be in inherited by the worker
+            _maybe_remove(sys.path, '')
+            _maybe_remove(sys.path, _mock_interactive_session_cwd)
+
+            # Add the desired session working directory
+            sys.path.insert(0, _mock_interactive_session_cwd)
+
+            with subprocess_worker(protocol=self.protocol) as w:
+                # Make the module unavailable in the remote worker
+                w.run(
+                    lambda p: sys.path.remove(p), _mock_interactive_session_cwd
+                )
+                # Import the actual file after starting the module since the
+                # worker is started using fork on Linux, which will inherits
+                # the parent sys.modules. On Python>3.6, the worker can be
+                # started using spawn using mp_context in ProcessPoolExectutor.
+                # TODO Once Python 3.6 reaches end of life, rely on mp_context
+                # instead.
+                import mock_local_folder.mod as mod
+                # The constructs whose pickling mechanism is changed using
+                # register_pickle_by_value are functions, classes, TypeVar and
+                # modules.
+                from mock_local_folder.mod import (
+                    local_function, LocalT, LocalClass
+                )
+
+                # Make sure the module/constructs are unimportable in the
+                # worker.
+                with pytest.raises(ImportError):
+                    w.run(lambda: __import__("mock_local_folder.mod"))
+                with pytest.raises(ImportError):
+                    w.run(
+                        lambda: __import__("mock_local_folder.subfolder.mod")
+                    )
+
+                for o in [mod, local_function, LocalT, LocalClass]:
+                    with pytest.raises(ImportError):
+                        w.run(lambda: o)
+
+                register_pickle_by_value(mod)
+                # function
+                assert w.run(lambda: local_function()) == local_function()
+                # typevar
+                assert w.run(lambda: LocalT.__name__) == LocalT.__name__
+                # classes
+                assert (
+                    w.run(lambda: LocalClass().method())
+                    == LocalClass().method()
+                )
+                # modules
+                assert (
+                    w.run(lambda: mod.local_function()) == local_function()
+                )
+
+                # Constructs from modules inside subfolders should be pickled
+                # by value if a namespace module pointing to some parent folder
+                # was registered for pickling by value. A "mock_local_folder"
+                # namespace module falls into that category, but a
+                # "mock_local_folder.mod" one does not.
+                from mock_local_folder.subfolder.submod import (
+                    LocalSubmodClass, LocalSubmodT, local_submod_function
+                )
+                # Shorter aliases to comply with line-length limits
+                _t, _func, _class = (
+                    LocalSubmodT, local_submod_function, LocalSubmodClass
+                )
+                with pytest.raises(ImportError):
+                    w.run(
+                        lambda: __import__("mock_local_folder.subfolder.mod")
+                    )
+                with pytest.raises(ImportError):
+                    w.run(lambda: local_submod_function)
+
+                unregister_pickle_by_value(mod)
+
+                with pytest.raises(ImportError):
+                    w.run(lambda: local_function)
+
+                with pytest.raises(ImportError):
+                    w.run(lambda: __import__("mock_local_folder.mod"))
+
+                # Test the namespace folder case
+                import mock_local_folder
+                register_pickle_by_value(mock_local_folder)
+                assert w.run(lambda: local_function()) == local_function()
+                assert w.run(lambda: _func()) == _func()
+                unregister_pickle_by_value(mock_local_folder)
+
+                with pytest.raises(ImportError):
+                    w.run(lambda: local_function)
+                with pytest.raises(ImportError):
+                    w.run(lambda: local_submod_function)
+
+                # Test the case of registering a single module inside a
+                # subfolder.
+                import mock_local_folder.subfolder.submod
+                register_pickle_by_value(mock_local_folder.subfolder.submod)
+                assert w.run(lambda: _func()) == _func()
+                assert w.run(lambda: _t.__name__) == _t.__name__
+                assert w.run(lambda: _class().method()) == _class().method()
+
+                # Registering a module from a subfolder for pickling by value
+                # should not make constructs from modules from the parent
+                # folder pickleable
+                with pytest.raises(ImportError):
+                    w.run(lambda: local_function)
+                with pytest.raises(ImportError):
+                    w.run(lambda: __import__("mock_local_folder.mod"))
+
+                unregister_pickle_by_value(
+                    mock_local_folder.subfolder.submod
+                )
+                with pytest.raises(ImportError):
+                    w.run(lambda: local_submod_function)
+
+                # Test the subfolder namespace module case
+                import mock_local_folder.subfolder
+                register_pickle_by_value(mock_local_folder.subfolder)
+                assert w.run(lambda: _func()) == _func()
+                assert w.run(lambda: _t.__name__) == _t.__name__
+                assert w.run(lambda: _class().method()) == _class().method()
+
+                unregister_pickle_by_value(mock_local_folder.subfolder)
+        finally:
+            _fname = "mock_local_folder"
+            sys.path = _prev_sys_path
+            for m in [_fname, f"{_fname}.mod", f"{_fname}.subfolder",
+                      f"{_fname}.subfolder.submod"]:
+                mod = sys.modules.pop(m, None)
+                if mod and mod.__name__ in list_registry_pickle_by_value():
+                    unregister_pickle_by_value(mod)
+
+    def test_pickle_constructs_from_installed_packages_registered_for_pickling_by_value(  # noqa
+        self
+    ):
+        for package_or_module in ["package", "module"]:
+            if package_or_module == "package":
+                import _cloudpickle_testpkg as m
+                f = m.package_function_with_global
+                _original_global = m.global_variable
+            elif package_or_module == "module":
+                import _cloudpickle_testpkg.mod as m
+                f = m.module_function_with_global
+                _original_global = m.global_variable
+            try:
+                with subprocess_worker(protocol=self.protocol) as w:
+                    assert w.run(lambda: f()) == _original_global
+
+                    # Test that f is pickled by value by modifying a global
+                    # variable that f uses, and making sure that this
+                    # modification shows up when calling the function remotely
+                    register_pickle_by_value(m)
+                    assert w.run(lambda: f()) == _original_global
+                    m.global_variable = "modified global"
+                    assert m.global_variable != _original_global
+                    assert w.run(lambda: f()) == "modified global"
+                    unregister_pickle_by_value(m)
+            finally:
+                m.global_variable = _original_global
+                if m.__name__ in list_registry_pickle_by_value():
+                    unregister_pickle_by_value(m)
+
+    def test_pickle_various_versions_of_the_same_function_with_different_pickling_method(  # noqa
+        self
+    ):
+        # Make sure that different versions of the same function (possibly
+        # pickled in a different way - by value and/or by reference) can
+        # peacefully co-exist (e.g. without globals interaction) in a remote
+        # worker.
+        import _cloudpickle_testpkg
+        from _cloudpickle_testpkg import package_function_with_global as f
+        _original_global = _cloudpickle_testpkg.global_variable
+
+        def _create_registry():
+            _main = __import__("sys").modules["__main__"]
+            _main._cloudpickle_registry = {}
+            # global _cloudpickle_registry
+
+        def _add_to_registry(v, k):
+            _main = __import__("sys").modules["__main__"]
+            _main._cloudpickle_registry[k] = v
+
+        def _call_from_registry(k):
+            _main = __import__("sys").modules["__main__"]
+            return _main._cloudpickle_registry[k]()
+
+        try:
+            with subprocess_worker(protocol=self.protocol) as w:
+                w.run(_create_registry)
+                w.run(_add_to_registry, f, "f_by_ref")
+
+                register_pickle_by_value(_cloudpickle_testpkg)
+                _cloudpickle_testpkg.global_variable = "modified global"
+                w.run(_add_to_registry, f, "f_by_val")
+                assert (
+                    w.run(_call_from_registry, "f_by_ref") == _original_global
+                )
+                assert (
+                    w.run(_call_from_registry, "f_by_val") == "modified global"
+                )
+
+        finally:
+            _cloudpickle_testpkg.global_variable = _original_global
+
+            if "_cloudpickle_testpkg" in list_registry_pickle_by_value():
+                unregister_pickle_by_value(_cloudpickle_testpkg)
+
     @pytest.mark.skipif(
         sys.version_info < (3, 7),
         reason="Determinism can only be guaranteed for Python 3.7+"
@@ -2406,6 +2639,45 @@ def test_lookup_module_and_qualname_stdlib_typevar():
     module, name = module_and_name
     assert module is typing
     assert name == 'AnyStr'
+
+
+def test_register_pickle_by_value():
+    import _cloudpickle_testpkg as pkg
+    import _cloudpickle_testpkg.mod as mod
+
+    assert list_registry_pickle_by_value() == set()
+
+    register_pickle_by_value(pkg)
+    assert list_registry_pickle_by_value() == {pkg.__name__}
+
+    register_pickle_by_value(mod)
+    assert list_registry_pickle_by_value() == {pkg.__name__, mod.__name__}
+
+    unregister_pickle_by_value(mod)
+    assert list_registry_pickle_by_value() == {pkg.__name__}
+
+    msg = f"Input should be a module object, got {pkg.__name__} instead"
+    with pytest.raises(ValueError, match=msg):
+        unregister_pickle_by_value(pkg.__name__)
+
+    unregister_pickle_by_value(pkg)
+    assert list_registry_pickle_by_value() == set()
+
+    msg = f"{pkg} is not registered for pickle by value"
+    with pytest.raises(ValueError, match=re.escape(msg)):
+        unregister_pickle_by_value(pkg)
+
+    msg = f"Input should be a module object, got {pkg.__name__} instead"
+    with pytest.raises(ValueError, match=msg):
+        register_pickle_by_value(pkg.__name__)
+
+    dynamic_mod = types.ModuleType('dynamic_mod')
+    msg = (
+        f"{dynamic_mod} was not imported correctly, have you used an "
+        f"`import` statement to access it?"
+    )
+    with pytest.raises(ValueError, match=re.escape(msg)):
+        register_pickle_by_value(dynamic_mod)
 
 
 def _all_types_to_test():