+++ CHANGES.md
@@ -6,9 +6,13 @@ dev
 
 - Python 3.5 is no longer supported.
 
+- Support for registering modules to be serialised by value. This will
+  allow for code defined in local modules to be serialised and executed
+  remotely without those local modules installed on the remote machine.
+  ([PR #417](https://github.com/cloudpipe/cloudpickle/pull/417))
+
 - Fix a side effect altering dynamic modules at pickling time.
   ([PR #426](https://github.com/cloudpipe/cloudpickle/pull/426))
-
 - Support for pickling type annotations on Python 3.10 as per [PEP 563](
   https://www.python.org/dev/peps/pep-0563/)
   ([PR #400](https://github.com/cloudpipe/cloudpickle/pull/400))